+++ falcon_auth2/utils/__init__.py
@@ -1,6 +1,6 @@
-from .asyncio_compat import await_
-from .asyncio_compat import greenlet_spawn
-from .classes import RequestAttributes
-from .functions import call_maybe_async
-from .functions import check_backend
-from .functions import check_getter
+from .asyncio_compat import await_ as await_
+from .asyncio_compat import greenlet_spawn as greenlet_spawn
+from .classes import RequestAttributes as RequestAttributes
+from .functions import call_maybe_async as call_maybe_async
+from .functions import check_backend as check_backend
+from .functions import check_getter as check_getter