+++ falcon_auth2/utils/asyncio_compat.py
@@ -1,40 +1,36 @@
 # based on sqlalchemy's lib/sqlalchemy/util/_concurrency_py3k.py
+from collections.abc import Callable
+from contextvars import copy_context
 import sys
 from typing import Any
-from typing import Callable
-from typing import Coroutine
+from typing import Awaitable
+from typing import NoReturn
+from typing import ParamSpec
+from typing import TypeVar
+
+_T = TypeVar("_T")
+_P = ParamSpec("_P")
 
 try:
     import greenlet
-
-    try:
-        from contextvars import copy_context as _copy_context
-
-        # If greenlet.gr_context is present in current version of greenlet,
-        # it will be set with a copy of the current context on creation.
-        # Refs: https://github.com/python-greenlet/greenlet/pull/198
-        getattr(greenlet.greenlet, "gr_context")
-    except (ImportError, AttributeError):
-        _copy_context = None
 
     # implementation based on snaury gist at
     # https://gist.github.com/snaury/202bf4f22c41ca34e56297bae5f33fef
     # Issue for context: https://github.com/python-greenlet/greenlet/issues/173
     class _AsyncIoGreenlet(greenlet.greenlet):
-        def __init__(self, fn, driver):
+        def __init__(self, fn: Callable[..., Any], driver: greenlet.greenlet | None) -> None:
             greenlet.greenlet.__init__(self, fn, driver)
             self.driver = driver
-            if _copy_context is not None:
-                self.gr_context = _copy_context()
+            self.gr_context = copy_context()
 
-    def await_(awaitable: Coroutine) -> Any:
+    def await_(awaitable: Awaitable[_T]) -> _T:
         """Awaits an async function in a sync method.
 
         The sync method must be insice a :func:`greenlet_spawn` context.
         :func:`await_` calls cannot be nested.
 
         Args:
-            awaitable (Coroutine): The coroutine to call.
+            awaitable (Awaitable): The awaitable to call.
 
         Raises:
             RuntimeError: If ``await_`` was called outside a :func:`greenlet_spawn` context or
@@ -54,9 +50,9 @@ try:
         # a coroutine to run. Once the awaitable is done, the driver greenlet
         # switches back to this greenlet with the result of awaitable that is
         # then returned to the caller (or raised as error)
-        return current.driver.switch(awaitable)
+        return current.driver.switch(awaitable)  # type: ignore[union-attr,no-any-return]
 
-    async def greenlet_spawn(fn: Callable, *args, **kwargs) -> Any:
+    async def greenlet_spawn(fn: Callable[_P, _T], *args: _P.args, **kwargs: _P.kwargs) -> _T:
         """Runs a sync function ``fn`` in a new greenlet.
 
         The sync function can then use :func:`await_` to wait for async functions.
@@ -91,20 +87,16 @@ try:
         finally:
             # clean up to avoid cycle resolution by gc
             del context.driver
-        return result
+        return result  # type: ignore[no-any-return]
 
 except ImportError:  # pragma: no cover
-    greenlet = None
+    greenlet = None  # type: ignore[assignment]
 
-    def _not_implemented():
-        # this conditional is to prevent pylance from considering
-        # greenlet_spawn() etc as "no return" and dimming out code below it
-        if greenlet:
-            return None
-        raise ValueError("Greenlet is required to use this function")
+    def _not_implemented() -> NoReturn:
+        raise ValueError("Greesnlet is required to use this function")
 
-    def await_(awaitable):
+    def await_(awaitable: Awaitable[_T]) -> _T:
         _not_implemented()
 
-    async def greenlet_spawn(fn, *args, **kw):
+    async def greenlet_spawn(fn: Callable[_P, _T], *args: _P.args, **kwargs: _P.kwargs) -> _T:
         _not_implemented()