+++ test/backends/test_jwt.py
@@ -1,8 +1,7 @@
 from datetime import datetime
 from datetime import timedelta
+from datetime import timezone
 
-from authlib.jose import JsonWebToken
-from authlib.jose import jwt
 import falcon
 import pytest
 
@@ -12,6 +11,11 @@ from falcon_auth2 import RequestAttributes
 from falcon_auth2.backends import JWTAuthBackend
 from .conftest import ConfigurableGetter
 from .conftest import ResourceFixture
+
+pytest.importorskip("authlib", reason="Authlib is required to run JWTAuthBackend tests")
+if True:  # avoid zimport reformatting these
+    from authlib.jose import JsonWebToken
+    from authlib.jose import jwt
 
 
 def jwt_token(key, payload, header=None, prefix="Bearer"):
@@ -27,6 +31,10 @@ def find_user(user_dict):
         return user_dict.get(payload["sub"])
 
     return m
+
+
+def utcnow():
+    return datetime.now(timezone.utc)
 
 
 class TestJWTAuth(ResourceFixture):
@@ -179,9 +187,9 @@ class TestJWTAuth(ResourceFixture):
             "iss": "my-iss",
             "sub": "1",
             "aud": "my-aud",
-            "exp": datetime.utcnow() + timedelta(seconds=10),
-            "nbf": datetime.utcnow(),
-            "iat": datetime.utcnow(),
+            "exp": utcnow() + timedelta(seconds=10),
+            "nbf": utcnow(),
+            "iat": utcnow(),
         }
         req = client.simulate_post("/auth", headers={"Authorization": jwt_token(key, payload)})
         assert req.text == str(user)
@@ -194,9 +202,9 @@ class TestJWTAuth(ResourceFixture):
                 "iss": "my-iss",
                 "sub": "1",
                 "aud": "my-aud",
-                "exp": datetime.utcnow() + timedelta(seconds=10),
-                "nbf": datetime.utcnow(),
-                "iat": datetime.utcnow(),
+                "exp": utcnow() + timedelta(seconds=10),
+                "nbf": utcnow(),
+                "iat": utcnow(),
             }
             del payload[key]
             req = client.simulate_post("/auth", headers={"Authorization": jwt_token(key, payload)})
@@ -208,9 +216,9 @@ class TestJWTAuth(ResourceFixture):
             "iss": "my-iss",
             "sub": "1",
             "aud": "my-aud",
-            "exp": datetime.utcnow() - timedelta(seconds=10),
-            "nbf": datetime.utcnow() - timedelta(seconds=10),
-            "iat": datetime.utcnow() - timedelta(seconds=10),
+            "exp": utcnow() - timedelta(seconds=10),
+            "nbf": utcnow() - timedelta(seconds=10),
+            "iat": utcnow() - timedelta(seconds=10),
         }
         req = client.simulate_post("/auth", headers={"Authorization": jwt_token(key, payload)})
         assert req.status == falcon.HTTP_UNAUTHORIZED