+++ test/test_middleware.py
@@ -1,4 +1,5 @@
 import falcon
+from falcon import App
 from falcon import testing
 import pytest
 
@@ -6,8 +7,6 @@ from falcon_auth2 import AuthMiddleware
 from falcon_auth2 import RequestAttributes
 from falcon_auth2.backends import AuthBackend
 from falcon_auth2.backends import NoAuthBackend
-from falcon_auth2.utils.compat import falcon2
-from .conftest import set_text
 from .conftest import User
 
 
@@ -22,10 +21,6 @@ def create_client(res, loader, mkw=None):
     mkw = mkw or {}
     backend = NoAuthBackend(loader)
     mkw.setdefault("backend", backend)
-    if falcon2:
-        from falcon import API as App
-    else:
-        from falcon import App
     app = App(middleware=AuthMiddleware(**mkw))
 
     app.add_route("/auth", res)
@@ -45,7 +40,7 @@ def auth(user, text="ok", backend=flag):
         assert req.context.auth["user"] is user
         if backend is not flag:
             assert req.context.auth["backend"] is backend
-        set_text(res, text)
+        res.text = text
 
     return m
 
@@ -53,7 +48,7 @@ def auth(user, text="ok", backend=flag):
 def no_auth(text):
     def m(req: falcon.Request, res, **kw):
         assert not hasattr(req.context, "auth")
-        set_text(res, text)
+        res.text = text
 
     return m
 
@@ -99,7 +94,7 @@ class TestAuthMiddleware:
             assert req.context.auth["user"] is user
             assert req.context.auth["backend"] is client.backend
 
-            set_text(res, "ok")
+            res.text = "ok"
 
         client = create_client(Resource(get), lambda auth: user)
 
@@ -142,7 +137,7 @@ class TestAuthMiddleware:
         def m(req, res):
             assert set(req.context.foobar.keys()) == {"user", "backend"}
             assert req.context.foobar["user"] is user
-            set_text(res, "ok")
+            res.text = "ok"
 
         client = create_client(Resource(m), lambda auth: user, dict(context_attr="foobar"))
 
@@ -180,7 +175,7 @@ class TestAuthMiddleware:
         def get(req, res):
             assert req.context.auth["user"] == 42
             assert req.context.auth["backend"] == "a-backend-here"
-            set_text(res, "ok")
+            res.text = "ok"
 
         client = create_client(Resource(get), lambda auth: 24, {"backend": MyBackend()})