+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0]
+- Ensure challenges are passed to single getter by ``MultiGetter``.
+- Dropped support for old python version. The library now requires python 3.10.
+- Fully typed the library.
+
 ## [0.1.0]
 - Add JWT support.